added watching of docker events and select some events to do a action
The actions we are interested in are only container related and those
events always have a 'type'.

To check if we will need to update the hosts file we only need the
follwoing statuses:

- start
- stop
- kill
- die

And therefore we will need to implement the actual handling and updating
of the hosts file.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>